+++ tests/unittest/_torch/auto_deploy/unit/singlegpu/custom_ops/test_trtllm_moe.py
@@ -1002,140 +1002,3 @@ def test_stack_nvfp4_moe_weights_transform_relu2(hidden_size, intermediate_size)
     # Should be close for FP4 quantization (gated MLP may have slightly larger diff due to alpha handling)
     tol = 1e-3
     torch.testing.assert_close(ref_output, transformed_output, rtol=tol, atol=tol)
-
-
-@pytest.mark.parametrize("allow_different_input_scales", [False, True])
-@pytest.mark.parametrize("scales_identical", [True, False])
-@pytest.mark.skipif(
-    not fp8_compatible() or not trtllm_ops_available(),
-    reason="Requires fp8 and trtllm support",
-)
-def test_fp8_moe_different_input_scales(allow_different_input_scales, scales_identical):
-    """
-    Test FP8 MoE behavior with different/identical input scales.
-
-    Tests the allow_different_input_scales config option:
-    - When scales_identical=True: should always work
-    - When scales_identical=False and allow_different_input_scales=False: should fail with assertion
-    - When scales_identical=False and allow_different_input_scales=True: should work (uses max)
-    """
-    import torch.fx as fx
-
-    from tensorrt_llm._torch.auto_deploy.transform.library.fused_moe import _stack_fp8_moe_weights
-
-    torch.manual_seed(0)
-
-    batch_size, num_experts, top_k = 4, 2, 2
-    hidden_size, intermediate_size = 128, 128
-    # Use non-gated MLP (Relu2) for this test
-    is_gated_mlp = False
-    act_fn = ActivationType.Relu2
-
-    # Generate test data
-    x = torch.randn(batch_size, hidden_size, dtype=torch.bfloat16, device="cuda") * 0.5
-    router_logits = torch.randn(batch_size, num_experts, dtype=torch.bfloat16, device="cuda")
-    routing_weights, selected_experts = compute_routing(router_logits, top_k)
-
-    # Create per-expert weights and scales (non-gated MLP, no w3)
-    # Use random FP8 data directly - we're testing transform logic, not quantization accuracy
-    w1_weight, w2_weight = [], []
-    w1_input_scale, w2_input_scale = [], []
-    w1_weight_scale, w2_weight_scale = [], []
-
-    for expert_id in range(num_experts):
-        # Random FP8 weights
-        w1_fp8 = torch.randn(intermediate_size, hidden_size, device="cuda").to(torch.float8_e4m3fn)
-        w2_fp8 = torch.randn(hidden_size, intermediate_size, device="cuda").to(torch.float8_e4m3fn)
-        w1_weight.append(w1_fp8)
-        w2_weight.append(w2_fp8)
-
-        # Random weight scales (shape [1])
-        w1_weight_scale.append(torch.tensor([0.1], dtype=torch.float32, device="cuda"))
-        w2_weight_scale.append(torch.tensor([0.1], dtype=torch.float32, device="cuda"))
-
-        # Input scales: either identical or different per expert (shape [1])
-        if scales_identical:
-            inp_scale = torch.tensor([1.0], dtype=torch.float32, device="cuda")
-        else:
-            # Different input scales per expert - big variance to test max() behavior
-            inp_scale = torch.tensor([0.5 + 0.5 * expert_id], dtype=torch.float32, device="cuda")
-
-        w1_input_scale.append(inp_scale)
-        w2_input_scale.append(inp_scale)
-
-    # Create a module with the FP8 MoE op
-    class FP8MoEModule(torch.nn.Module):
-        def __init__(self):
-            super().__init__()
-            for i in range(num_experts):
-                self.register_buffer(f"w1_{i}", w1_weight[i])
-                self.register_buffer(f"w2_{i}", w2_weight[i])
-                self.register_buffer(f"w1_iscale_{i}", w1_input_scale[i])
-                self.register_buffer(f"w2_iscale_{i}", w2_input_scale[i])
-                self.register_buffer(f"w1_wscale_{i}", w1_weight_scale[i])
-                self.register_buffer(f"w2_wscale_{i}", w2_weight_scale[i])
-
-        def forward(self, x, selected_experts, routing_weights):
-            return torch.ops.auto_deploy.torch_quant_fp8_moe(
-                x,
-                selected_experts,
-                routing_weights,
-                [getattr(self, f"w1_{i}") for i in range(num_experts)],
-                [getattr(self, f"w2_{i}") for i in range(num_experts)],
-                [],  # w3 is empty for non-gated MLP
-                [getattr(self, f"w1_iscale_{i}") for i in range(num_experts)],
-                [getattr(self, f"w2_iscale_{i}") for i in range(num_experts)],
-                [],  # w3 input scale is empty for non-gated MLP
-                [getattr(self, f"w1_wscale_{i}") for i in range(num_experts)],
-                [getattr(self, f"w2_wscale_{i}") for i in range(num_experts)],
-                [],  # w3 weight scale is empty for non-gated MLP
-                is_gated_mlp=is_gated_mlp,
-                act_fn=act_fn,
-            )
-
-    module = FP8MoEModule().cuda()
-    gm = fx.symbolic_trace(module)
-
-    # Compute reference output from original graph before transformation
-    with torch.inference_mode():
-        ref_output = gm(x, selected_experts, routing_weights)
-
-    # Expected behavior:
-    # - scales_identical=True: always works
-    # - scales_identical=False, allow_different_input_scales=False: assertion error
-    # - scales_identical=False, allow_different_input_scales=True: works with max()
-
-    if not scales_identical and not allow_different_input_scales:
-        # Should fail with assertion error
-        with pytest.raises(AssertionError, match="input scales should have the same value"):
-            _stack_fp8_moe_weights(
-                gm, backend="trtllm", allow_different_input_scales=allow_different_input_scales
-            )
-    else:
-        # Should succeed
-        num_transformed = _stack_fp8_moe_weights(
-            gm, backend="trtllm", allow_different_input_scales=allow_different_input_scales
-        )
-        gm.recompile()
-
-        assert num_transformed == 1, f"Expected 1 transform, got {num_transformed}"
-
-        # Verify that max() is used when scales differ
-        if not scales_identical:
-            expected_max_w1_input_scale = torch.stack(w1_input_scale).max()
-            actual_w1_input_max = getattr(gm, "quant_moe_fc1_act_scale_max_0")
-
-            assert torch.allclose(actual_w1_input_max, expected_max_w1_input_scale), (
-                f"w1 input scale max mismatch. Got {actual_w1_input_max}, expected {expected_max_w1_input_scale}"
-            )
-
-        # Run the transformed graph and compare to reference output
-        with torch.inference_mode():
-            output = gm(x, selected_experts, routing_weights)
-            assert output.shape == ref_output.shape, (
-                f"Output shape mismatch: {output.shape} vs {ref_output.shape}"
-            )
-
-            assert torch.allclose(output, ref_output, rtol=0.05, atol=0.05), (
-                f"Output mismatch. rtol=0.1, atol=0.1Max diff: {(output - ref_output).abs().max()}"
-            )
